0x13-more_singly_linked_lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -9,7 +9,7 @@
 size_t print_listint(const listint_t *header)
 {
 	const listint_t *iTemp;
-	unsigned int iCount = 0;
+	size_t iCount = 0;
 
 	iTemp = header;
 	while (iTemp)
